@@ -46,7 +46,14 @@ func (app *App) initStores(ctx context.Context) error {
 		fallback.Scheme = "http"
 		fallback.Host = app.l.Addr().String()
 		fallback.Path = app.cfg.HTTPPrefix
-		app.ConfigStore, err = config.NewStore(ctx, app.db, app.cfg.EncryptionKeys, app.cfg.PublicURL, fallback.String())
+		storeCfg := config.StoreConfig{
+			DB:                 app.db,
+			Keys:               app.cfg.EncryptionKeys,
+			FallbackURL:        fallback.String(),
+			ExplicitURL:        app.cfg.PublicURL,
+			IngressEmailDomain: app.cfg.EmailIntegrationDomain,
+		}
+		app.ConfigStore, err = config.NewStore(ctx, storeCfg)
 	}
 	if err != nil {
 		return errors.Wrap(err, "init config store")
