@@ -25,35 +25,52 @@ import (
 
 // Store handles saving and loading configuration from a postgres database.
 type Store struct {
-	rawCfg       Config
-	cfgVers      int
-	fallbackURL  string
-	explicitURL  string
-	mx           sync.RWMutex
-	db           *sql.DB
-	keys         keyring.Keys
-	latestConfig *sql.Stmt
-	setConfig    *sql.Stmt
-	lock         *sql.Stmt
+	rawCfg             Config
+	cfgVers            int
+	fallbackURL        string
+	explicitURL        string
+	ingressEmailDomain string
+	mx                 sync.RWMutex
+	db                 *sql.DB
+	keys               keyring.Keys
+	latestConfig       *sql.Stmt
+	setConfig          *sql.Stmt
+	lock               *sql.Stmt
 
 	closeCh chan struct{}
 }
 
-// NewStore will create a new Store with the given parameters. It will automatically detect
+type StoreConfig struct {
+	DB   *sql.DB
+	Keys keyring.Keys
+
+	// FallbackURL is the URL to use when the DB config does not specify a public URL.
+	FallbackURL string
+
+	// ExplicitURL is the full public URL to use for all links.
+	ExplicitURL string
+
+	// IngressEmailDomain is the domain to use for ingress email addresses.
+	IngressEmailDomain string
+}
+
+// NewStore will create a new Store with the given StoreConfig parameters. It will automatically detect
 // new configuration changes.
-func NewStore(ctx context.Context, db *sql.DB, keys keyring.Keys, explicitURL, fallbackURL string) (*Store, error) {
-	p := util.Prepare{Ctx: ctx, DB: db}
+func NewStore(ctx context.Context, cfg StoreConfig) (*Store, error) {
+	p := util.Prepare{Ctx: ctx, DB: cfg.DB}
 
 	s := &Store{
-		db:           db,
-		fallbackURL:  fallbackURL,
-		explicitURL:  explicitURL,
-		latestConfig: p.P(`select id, data, schema from config where schema <= $1 order by id desc limit 1`),
-		setConfig:    p.P(`insert into config (id, schema, data) values (DEFAULT, $1, $2) returning (id)`),
-		lock:         p.P(`lock config in exclusive mode`),
-		keys:         keys,
-		closeCh:      make(chan struct{}),
+		db:                 cfg.DB,
+		fallbackURL:        cfg.FallbackURL,
+		explicitURL:        cfg.ExplicitURL,
+		ingressEmailDomain: cfg.IngressEmailDomain,
+		latestConfig:       p.P(`select id, data, schema from config where schema <= $1 order by id desc limit 1`),
+		setConfig:          p.P(`insert into config (id, schema, data) values (DEFAULT, $1, $2) returning (id)`),
+		lock:               p.P(`lock config in exclusive mode`),
+		keys:               cfg.Keys,
+		closeCh:            make(chan struct{}),
 	}
+
 	if p.Err != nil {
 		return nil, p.Err
 	}
@@ -128,6 +145,7 @@ func (s *Store) Reload(ctx context.Context) error {
 	rawCfg := *cfg
 	rawCfg.fallbackURL = s.fallbackURL
 	rawCfg.explicitURL = s.explicitURL
+	rawCfg.intEmailDomain = s.ingressEmailDomain
 
 	err = cfg.Validate()
 	if err != nil {
@@ -139,7 +157,6 @@ func (s *Store) Reload(ctx context.Context) error {
 	s.cfgVers = id
 	s.rawCfg = rawCfg
 	s.mx.Unlock()
-
 	if oldVers != id {
 		log.Logf(ctx, "Loaded config version %d ", id)
 	}
