@@ -110,7 +110,7 @@ func (l *Loader[K, V]) load(entries []*entry[K, V]) []*entry[K, V] {
 		// Call fetch for everything we're loading
 		res, err := l.cfg.FetchFunc(ctx, ids)
 		if err != nil {
-			// If the fetch failed, set all the pending entires err property to
+			// If the fetch failed, set all the pending entries err property to
 			// reflect the failure, and close the done channel to indicate the load/fetch
 			// completed.
 			for _, e := range cpy {
@@ -132,7 +132,7 @@ func (l *Loader[K, V]) load(entries []*entry[K, V]) []*entry[K, V] {
 			e.data = &res[i]
 		}
 
-		// nil or not, all entires are now done loading, if the .data prop was not set
+		// nil or not, all entries are now done loading, if the .data prop was not set
 		// then the entry does not exist.
 		for _, e := range cpy {
 			close(e.done)
