@@ -11,7 +11,7 @@ import (
 func (a *Query) IntegrationKeyTypes(ctx context.Context) ([]graphql2.IntegrationKeyTypeInfo, error) {
 	cfg := config.FromContext(ctx)
 	return []graphql2.IntegrationKeyTypeInfo{
-		{ID: "email", Name: "Email", Label: "Email Address", Enabled: cfg.Mailgun.Enable},
+		{ID: "email", Name: "Email", Label: "Email Address", Enabled: cfg.EmailIngressEnabled()},
 		{ID: "generic", Name: "Generic API", Label: "Generic Webhook URL", Enabled: true},
 		{ID: "grafana", Name: "Grafana", Label: "Grafana Webhook URL", Enabled: true},
 		{ID: "site24x7", Name: "Generic", Label: "Site24x7 Webhook URL", Enabled: true},
