@@ -56,7 +56,7 @@ var searchTemplate = template.Must(template.New("search").Funcs(search.Helpers()
 		AND NOT sched.id = any(:omit)
 	{{end}}
 	{{if .Search}}
-		AND {{prefixSearch "search" "sched.name"}}
+		AND {{orderedPrefixSearch "search" "sched.name"}}
 	{{end}}
 	{{if .After.Name}}
 		AND
