@@ -77,7 +77,7 @@ var searchTemplate = template.Must(template.New("search").Funcs(search.Helpers()
 		)
 	{{end}}
 	{{- if and .Search (not .LabelKey) (not .IntegrationKey)}}
-		AND {{prefixSearch "search" "svc.name"}}
+		AND {{orderedPrefixSearch "search" "svc.name"}}
 	{{- end}}
 	{{- if .After.Name}}
 		AND
