@@ -2,7 +2,7 @@ import { test } from '@playwright/test'
 
 test.describe.configure({ mode: 'parallel' })
 
-// test loging in with OIDC
+// test logging in with OIDC
 test('OIDC Login', async ({ page }) => {
   await page.goto('./profile')
 
@@ -11,7 +11,7 @@ test('OIDC Login', async ({ page }) => {
   // ensure we have an h1 with jane.doe
   await page.waitForSelector('h1 >> "jane.doe"')
 })
-// test loging in with OIDC
+// test logging in with OIDC
 test('OIDC Login (fallback public url)', async ({ page }) => {
   await page.goto('http://127.0.0.1:6120/profile')
 
