@@ -42,8 +42,12 @@ export default function SWONode({ node, name }: SWONodeProps): JSX.Element {
             </ListItem>
             <ListItem>
               <ListItemText primary='Connections' />
-              <ListItemSecondaryAction>
-                {node.connections?.reduce((acc, cur) => acc + cur.count, 0)}
+              <ListItemSecondaryAction
+                sx={{ width: '24px', textAlign: 'center' }}
+              >
+                <Typography>
+                  {node.connections?.reduce((acc, cur) => acc + cur.count, 0)}
+                </Typography>
               </ListItemSecondaryAction>
             </ListItem>
           </List>
@@ -91,8 +95,12 @@ export default function SWONode({ node, name }: SWONodeProps): JSX.Element {
           </ListItem>
           <ListItem>
             <ListItemText primary='Connections' />
-            <ListItemSecondaryAction>
-              {node.connections?.reduce((acc, cur) => acc + cur.count, 0)}
+            <ListItemSecondaryAction
+              sx={{ width: '24px', textAlign: 'center' }}
+            >
+              <Typography>
+                {node.connections?.reduce((acc, cur) => acc + cur.count, 0)}
+              </Typography>
             </ListItemSecondaryAction>
           </ListItem>
         </List>
