@@ -4,10 +4,10 @@ let n = 1
 let u = 1
 const names: { [key: string]: string } = {}
 
-// friendlyName will assign a persistant "friendly" name to the node.
+// friendlyName will assign a persistent "friendly" name to the node.
 //
 // This ensures a specific ID will always refer to the same node. This
-// is so that it is clear if a node dissapears or a new one appears.
+// is so that it is clear if a node disappears or a new one appears.
 //
 // Note: `Node 1` on one browser tab may not be the same node as `Node 1`
 // on another browser tab.
