@@ -253,9 +253,11 @@ export default function ScheduleCalendarEventWrapper({
           vertical: 'top',
           horizontal: 'left',
         }}
-        PaperProps={{
-          // @ts-expect-error - DOM attr for tests
-          'data-cy': 'shift-tooltip',
+        slotProps={{
+          paper: {
+            // @ts-expect-error - DOM attr for tests
+            'data-cy': 'shift-tooltip',
+          },
         }}
         classes={{
           paper: classes.paper,
