@@ -1,6 +1,5 @@
 import { gql, QueryResult, useQuery } from '@apollo/client'
 import { DateTime } from 'luxon'
-import { Schedule } from '../../schema'
 
 const schedTZQuery = gql`
   query ($id: ID!) {
@@ -13,7 +12,7 @@ const schedTZQuery = gql`
 
 interface ScheduleTZResult {
   // q is the Apollo query status
-  q: QueryResult<Schedule, { id: string }>
+  q: QueryResult
   // zone is schedule time zone name if ready; else empty string
   zone: string
   // isLocalZone is true if schedule and system time zone are equal
